- правки

+++ src/Game/EditorConstant.h
@@ -2,7 +2,7 @@
 
 const float CameraSpeedInEditorLeftPanel = 50.0f;
 
-const float gridSize = 1000.0f;
-const float viewSize = 100.0f;
-const float gridStep = 3.0f;
+const int gridSize = 1000;
+const int viewSize = 100;
+const int gridStep = 3;
 const float gridLineSize = 0.0003f;
